+++ lib/commercetools-api/src/Models/Me/MyTransactionDraftBuilder.php
@@ -56,6 +56,12 @@ final class MyTransactionDraftBuilder implements Builder
     private $custom;
 
     /**
+
+     * @var ?string
+     */
+    private $interfaceId;
+
+    /**
      * <p>Date and time (UTC) the Transaction took place.</p>
      *
 
@@ -110,6 +116,17 @@ final class MyTransactionDraftBuilder implements Builder
     public function getCustom()
     {
         return $this->custom instanceof CustomFieldsDraftBuilder ? $this->custom->build() : $this->custom;
+    }
+
+    /**
+     * <p>Identifier used by the payment service that processes the Payment (for example, a PSP) in the current transaction.</p>
+     *
+
+     * @return null|string
+     */
+    public function getInterfaceId()
+    {
+        return $this->interfaceId;
     }
 
     /**
@@ -168,6 +185,17 @@ final class MyTransactionDraftBuilder implements Builder
     }
 
     /**
+     * @param ?string $interfaceId
+     * @return $this
+     */
+    public function withInterfaceId(?string $interfaceId)
+    {
+        $this->interfaceId = $interfaceId;
+
+        return $this;
+    }
+
+    /**
      * @deprecated use withAmount() instead
      * @return $this
      */
@@ -196,7 +224,8 @@ final class MyTransactionDraftBuilder implements Builder
             $this->type,
             $this->amount instanceof MoneyBuilder ? $this->amount->build() : $this->amount,
             $this->interactionId,
-            $this->custom instanceof CustomFieldsDraftBuilder ? $this->custom->build() : $this->custom
+            $this->custom instanceof CustomFieldsDraftBuilder ? $this->custom->build() : $this->custom,
+            $this->interfaceId
         );
     }