+++ lib/commercetools-api/src/Models/Me/MyTransactionDraftModel.php
@@ -54,6 +54,12 @@ final class MyTransactionDraftModel extends JsonObjectModel implements MyTransac
      */
     protected $custom;
 
+    /**
+     *
+     * @var ?string
+     */
+    protected $interfaceId;
+
 
     /**
      * @psalm-suppress MissingParamType
@@ -63,13 +69,15 @@ final class MyTransactionDraftModel extends JsonObjectModel implements MyTransac
         ?string $type = null,
         ?Money $amount = null,
         ?string $interactionId = null,
-        ?CustomFieldsDraft $custom = null
+        ?CustomFieldsDraft $custom = null,
+        ?string $interfaceId = null
     ) {
         $this->timestamp = $timestamp;
         $this->type = $type;
         $this->amount = $amount;
         $this->interactionId = $interactionId;
         $this->custom = $custom;
+        $this->interfaceId = $interfaceId;
     }
 
     /**
@@ -180,6 +188,26 @@ final class MyTransactionDraftModel extends JsonObjectModel implements MyTransac
         return $this->custom;
     }
 
+    /**
+     * <p>Identifier used by the payment service that processes the Payment (for example, a PSP) in the current transaction.</p>
+     *
+     *
+     * @return null|string
+     */
+    public function getInterfaceId()
+    {
+        if (is_null($this->interfaceId)) {
+            /** @psalm-var ?string $data */
+            $data = $this->raw(self::FIELD_INTERFACE_ID);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->interfaceId = (string) $data;
+        }
+
+        return $this->interfaceId;
+    }
+
 
     /**
      * @param ?DateTimeImmutable $timestamp
@@ -219,6 +247,14 @@ final class MyTransactionDraftModel extends JsonObjectModel implements MyTransac
     public function setCustom(?CustomFieldsDraft $custom): void
     {
         $this->custom = $custom;
+    }
+
+    /**
+     * @param ?string $interfaceId
+     */
+    public function setInterfaceId(?string $interfaceId): void
+    {
+        $this->interfaceId = $interfaceId;
     }