+++ lib/commercetools-api/src/Models/Payment/PaymentSetTransactionInterfaceIdActionModel.php
@@ -1,0 +1,130 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Api\Models\Payment;
+
+use Commercetools\Base\DateTimeImmutableCollection;
+use Commercetools\Base\JsonObject;
+use Commercetools\Base\JsonObjectModel;
+use Commercetools\Base\MapperFactory;
+use stdClass;
+
+/**
+ * @internal
+ */
+final class PaymentSetTransactionInterfaceIdActionModel extends JsonObjectModel implements PaymentSetTransactionInterfaceIdAction
+{
+    public const DISCRIMINATOR_VALUE = 'setTransactionInterfaceId';
+    /**
+     *
+     * @var ?string
+     */
+    protected $action;
+
+    /**
+     *
+     * @var ?string
+     */
+    protected $transactionId;
+
+    /**
+     *
+     * @var ?string
+     */
+    protected $interfaceId;
+
+
+    /**
+     * @psalm-suppress MissingParamType
+     */
+    public function __construct(
+        ?string $transactionId = null,
+        ?string $interfaceId = null,
+        ?string $action = null
+    ) {
+        $this->transactionId = $transactionId;
+        $this->interfaceId = $interfaceId;
+        $this->action = $action ?? self::DISCRIMINATOR_VALUE;
+    }
+
+    /**
+     *
+     * @return null|string
+     */
+    public function getAction()
+    {
+        if (is_null($this->action)) {
+            /** @psalm-var ?string $data */
+            $data = $this->raw(self::FIELD_ACTION);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->action = (string) $data;
+        }
+
+        return $this->action;
+    }
+
+    /**
+     * <p>Unique identifier of the <a href="ctp:api:type:Transaction">Transaction</a>.</p>
+     *
+     *
+     * @return null|string
+     */
+    public function getTransactionId()
+    {
+        if (is_null($this->transactionId)) {
+            /** @psalm-var ?string $data */
+            $data = $this->raw(self::FIELD_TRANSACTION_ID);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->transactionId = (string) $data;
+        }
+
+        return $this->transactionId;
+    }
+
+    /**
+     * <p>Identifier used by the payment service that processes the Payment (for example, a PSP) in the current transaction. It must be unique across all transactions.
+     * If <code>interfaceId</code> is absent, this field will be removed from the specified Transaction, if it exists.</p>
+     *
+     *
+     * @return null|string
+     */
+    public function getInterfaceId()
+    {
+        if (is_null($this->interfaceId)) {
+            /** @psalm-var ?string $data */
+            $data = $this->raw(self::FIELD_INTERFACE_ID);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->interfaceId = (string) $data;
+        }
+
+        return $this->interfaceId;
+    }
+
+
+    /**
+     * @param ?string $transactionId
+     */
+    public function setTransactionId(?string $transactionId): void
+    {
+        $this->transactionId = $transactionId;
+    }
+
+    /**
+     * @param ?string $interfaceId
+     */
+    public function setInterfaceId(?string $interfaceId): void
+    {
+        $this->interfaceId = $interfaceId;
+    }
+}