+++ lib/commercetools-api/src/Models/Payment/TransactionDraft.php
@@ -22,6 +22,7 @@ interface TransactionDraft extends JsonObject
     public const FIELD_INTERACTION_ID = 'interactionId';
     public const FIELD_STATE = 'state';
     public const FIELD_CUSTOM = 'custom';
+    public const FIELD_INTERFACE_ID = 'interfaceId';
 
     /**
      * <p>Date and time (UTC) the Transaction took place.</p>
@@ -73,6 +74,14 @@ interface TransactionDraft extends JsonObject
     public function getCustom();
 
     /**
+     * <p>Identifier used by the payment service that processes the Payment (for example, a PSP) in the current transaction.</p>
+     *
+
+     * @return null|string
+     */
+    public function getInterfaceId();
+
+    /**
      * @param ?DateTimeImmutable $timestamp
      */
     public function setTimestamp(?DateTimeImmutable $timestamp): void;
@@ -101,4 +110,9 @@ interface TransactionDraft extends JsonObject
      * @param ?CustomFieldsDraft $custom
      */
     public function setCustom(?CustomFieldsDraft $custom): void;
+
+    /**
+     * @param ?string $interfaceId
+     */
+    public function setInterfaceId(?string $interfaceId): void;
 }